Move unit tests that were actually e2e to e2e tests
