+++ tests/unit_tests/sync/test_substrate_interface_sync_unit.py
@@ -1,4 +1,3 @@
-import tracemalloc
 from unittest.mock import MagicMock
 
 from async_substrate_interface.sync_substrate import (
@@ -6,7 +5,6 @@ from async_substrate_interface.sync_substrate import (
     QueryMapResult,
     ExtrinsicReceipt,
 )
-from tests.helpers.settings import ARCHIVE_ENTRYPOINT, LATENT_LITE_ENTRYPOINT
 
 
 def test_runtime_call(monkeypatch):
@@ -60,50 +58,6 @@ def test_runtime_call(monkeypatch):
     )
     substrate.close()
     print("test_runtime_call succeeded")
-
-
-def test_runtime_switching():
-    print("Testing test_runtime_switching")
-    block = 6067945  # block where a runtime switch happens
-    with SubstrateInterface(
-        ARCHIVE_ENTRYPOINT, ss58_format=42, chain_name="Bittensor"
-    ) as substrate:
-        # assures we switch between the runtimes without error
-        assert substrate.get_extrinsics(block_number=block - 20) is not None
-        assert substrate.get_extrinsics(block_number=block) is not None
-        assert substrate.get_extrinsics(block_number=block - 21) is not None
-    print("test_runtime_switching succeeded")
-
-
-def test_memory_leak():
-    import gc
-
-    # Stop any existing tracemalloc and start fresh
-    tracemalloc.stop()
-    tracemalloc.start()
-    two_mb = 2 * 1024 * 1024
-
-    # Warmup: populate caches before taking baseline
-    for _ in range(2):
-        subtensor = SubstrateInterface(LATENT_LITE_ENTRYPOINT)
-        subtensor.close()
-
-    baseline_snapshot = tracemalloc.take_snapshot()
-
-    for i in range(5):
-        subtensor = SubstrateInterface(LATENT_LITE_ENTRYPOINT)
-        subtensor.close()
-        gc.collect()
-
-        snapshot = tracemalloc.take_snapshot()
-        stats = snapshot.compare_to(baseline_snapshot, "lineno")
-        total_diff = sum(stat.size_diff for stat in stats)
-        current, peak = tracemalloc.get_traced_memory()
-        # Allow cumulative growth up to 2MB per iteration from baseline
-        assert total_diff < two_mb * (i + 1), (
-            f"Loop {i}: diff={total_diff / 1024:.2f} KiB, current={current / 1024:.2f} KiB, "
-            f"peak={peak / 1024:.2f} KiB"
-        )
 
 
 def test_async_query_map_result_retrieve_all_records():
